requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
certifi==2018.4.16
chardet==3.0.4
click==6.7
dnspython==1.15.0
first==2.0.1
idna==2.6
pip-tools==2.0.2
requests==2.18.4
six==1.11.0
urllib3==1.22
